extracted parsing and selection to util files
